\name{dataSets}
\docType{data}
\alias{datacf}
\alias{dataIn}
\alias{dataOut1}
\alias{dataOut2}
\title{Two artificial data sets}

\description{
	The first artificial dataset is generated by the function :

	f1 = function(t)

	\{ return( 0.8*atan(a*t) + exp(b*(-4-t)+1) + log(c*(4-t)+1) ) \}

	and the second one by :

	f2 = function(t)

	\{ return( 0.8*atan(a*(4-t)) + exp(b*(-4+t)+1) + log(c*(4+t)+1) ) \}

	for a, b, c varying uniformly in [0,4] or [0,7].

	The matrix \code{dataIn} contains the input parameters a, b, c in rows, while the matrices \code{dataOut1} and \code{dataOut2} are filled with the corresponding curves in rows (200 sample points).
}

\usage{datacf}

\format{Matrices with 300 rows, and 3 columns for \code{dataIn}, 200 for the others (sample points).}

