\name{Z_modelcf}
\alias{modelcf}
\title{package modelcf}

\description{
	This package contains a generic way to build surrogate models of physical computer codes, when inputs are vectors (in R^p) and outputs (continuous) curves from [a,b] to R. The curves are discretized on a finite grid t1,...,tD.
}

\seealso{
	\code{\link{fmetam}}, \code{\link{predict.modelcf}}, \code{\link{mixpredf}}, \code{\link{nfoldcv}}.
}

