\name{modelcf}
\alias{modelcf}
\title{package modelcf}

\description{
	This package contains a generic way to build surrogate models of physical computer codes, when inputs are vectors (in R^p) and outputs (continuous) curves from [a,b] to R. The curves are discretized on a finite grid t1,...,tD.
}

\seealso{
	\code{\link{fmetam}}, \code{\link{predict.modelcf}}, \code{\link{mixpredf}}, \code{\link{nfoldcv}}.
}

\note{
	This work was supported by the CEA Cadarache, where I was employed three years for my PhD.
	I would also like to thank Bertrand Iooss and Michel Marques for all their suggestions of improvement,
	and all the people who helped me, directly or not, for creating this package.
}

